docker_compose/simple_project/app/requirements.txt
django==4.0.4
uwsgi==2.0.20
python-dotenv==0.21.0
psycopg2-binary==2.9.5
django-split-settings==1.2.0
djangorestframework
